wip: attestation fr and ecdsa attestations
