+++ circuit/Cargo.toml
@@ -18,6 +18,7 @@ num-traits = "0.2.11"
 num-rational = "0.4.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+secp256k1 = { version = "0.27.0", features = ["recovery"] }
 
 [dependencies.snark-verifier]
 git = "https://github.com/privacy-scaling-explorations/snark-verifier"
