+++ client/Cargo.toml
@@ -15,5 +15,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 csv = "1.1"
 once_cell = "1.17.1"
+revm-primitives = "1.1.2"
 
 eigen-trust-circuit = { path = "../circuit" }
